\name{jterator-package}
\alias{jterator-package}
\alias{jterator}
\docType{package}
\title{
Input/output for jterator module
}
\description{
Reads module description from handles file (YAML) and reads and writes data in HDF5 file.
}
\details{
\tabular{ll}{
Package: \tab jterator\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-07\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Markus Herrmann
Maintainer: Markus Herrmann <markus.herrmann@imls.uzh.ch>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
