% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jterator.R
\name{writedata}
\alias{writedata}
\title{Jterator API.
Writing data to HDF5 file using the location specified in "handles".}
\usage{
writedata(handles, data)
}
\arguments{
\item{handles}{List of module input/output information.}

\item{List}{of data output arguments that should be written to HDF5 file.}
}
\description{
Jterator API.
Writing data to HDF5 file using the location specified in "handles".
}

