% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jterator.R
\name{writeoutputargs}
\alias{writeoutputargs}
\title{Jterator API.
Writing output arguments to HDF5 file using the location specified in "handles".}
\usage{
writeoutputargs(handles, output_args)
}
\arguments{
\item{handles}{List of module input/output information.}

\item{List}{of temporary pipeline output arguments that should be written to HDF5 file.}
}
\description{
Jterator API.
Writing output arguments to HDF5 file using the location specified in "handles".
}

